# Add `shuffle!` algorithm specialized for `AbstractArray{Bool}`