Why? This is faster.

Benchmarks using my new, in development benchmarking package. They only take 50s to reproduce (excluding time to install `Chairmarks`) and are pastable into a REPL.

```julia
using Random
import Pkg; Pkg.add(url="https://github.com/LilithHafner/Chairmarks.jl", rev="bbef8833c0a9ecb647181fbfbded0757ab52d094") # v0.1.0
using Chairmarks

function my_shuffle!(v::AbstractVector{Bool})
    old_sum = sum(v)
    x = 2old_sum <= length(v)
    fuel = x ? old_sum : length(v) - old_sum
    fuel == 0 && return v
    v .= !x
    while 0 < fuel
        k = rand(eachindex(v))
        fuel -= v[k] != x
        v[k] = x
    end
    v
end

@time for n in [0,1,2,3,10,30,100,300,1000,3000,10_000,100_000,10_000_000,100_000_000]
    for k in (n == 0 ? [0] : unique([0, 1, div(n,3), div(n,2), 2div(n, 3), n-1, n]))
        println("$k/$n")
        for T in [Vector, BitVector]
            print(lpad(T, 9), "  ")
            old = @b T(vcat(trues(k), falses(n-k))) shuffle! sum(_) == k || error()
            new = @b T(vcat(trues(k), falses(n-k))) my_shuffle! sum(_) == k || error()
            show(stdout, MIME"text/plain"(), old)
            print(" => ")
            show(stdout, MIME"text/plain"(), new)
            println(old.time < new.time ? " REGRESSION" : "")
        end
    end
end
```

Results
```julia
0/0
   Vector  2.918 ns => 3.513 ns REGRESSION
BitVector  2.921 ns => 2.841 ns
0/1
   Vector  2.845 ns => 3.766 ns REGRESSION
BitVector  2.924 ns => 2.916 ns
1/1
   Vector  2.919 ns => 3.751 ns REGRESSION
BitVector  2.919 ns => 2.918 ns
0/2
   Vector  10.879 ns => 4.100 ns
BitVector  10.506 ns => 2.917 ns
1/2
   Vector  10.812 ns => 7.024 ns
BitVector  10.492 ns => 6.630 ns
2/2
   Vector  10.591 ns => 4.098 ns
BitVector  10.626 ns => 2.919 ns
0/3
   Vector  13.992 ns => 4.948 ns
BitVector  15.341 ns => 2.920 ns
1/3
   Vector  14.229 ns => 9.300 ns
BitVector  15.378 ns => 6.634 ns
2/3
   Vector  14.221 ns => 9.295 ns
BitVector  15.410 ns => 7.892 ns
3/3
   Vector  14.140 ns => 4.946 ns
BitVector  15.263 ns => 2.920 ns
0/10
   Vector  48.060 ns => 4.684 ns
BitVector  51.525 ns => 2.916 ns
1/10
   Vector  48.167 ns => 9.160 ns
BitVector  51.877 ns => 6.633 ns
3/10
   Vector  48.625 ns => 16.276 ns
BitVector  51.923 ns => 15.691 ns
5/10
   Vector  48.460 ns => 33.958 ns
BitVector  51.348 ns => 33.815 ns
6/10
   Vector  48.343 ns => 25.318 ns
BitVector  51.618 ns => 24.984 ns
9/10
   Vector  47.867 ns => 11.040 ns
BitVector  51.661 ns => 7.906 ns
10/10
   Vector  48.227 ns => 4.683 ns
BitVector  51.264 ns => 2.920 ns
0/30
   Vector  142.733 ns => 7.607 ns
BitVector  153.792 ns => 2.919 ns
1/30
   Vector  142.578 ns => 14.358 ns
BitVector  155.433 ns => 6.627 ns
10/30
   Vector  143.401 ns => 55.949 ns
BitVector  155.714 ns => 51.909 ns
15/30
   Vector  143.633 ns => 80.332 ns
BitVector  155.799 ns => 78.625 ns
20/30
   Vector  143.216 ns => 55.852 ns
BitVector  155.120 ns => 52.758 ns
29/30
   Vector  142.465 ns => 14.345 ns
BitVector  156.191 ns => 7.910 ns
30/30
   Vector  142.326 ns => 7.615 ns
BitVector  156.197 ns => 2.920 ns
0/100
   Vector  523.585 ns => 7.606 ns
BitVector  565.000 ns => 2.920 ns
1/100
   Vector  529.315 ns => 17.129 ns
BitVector  567.500 ns => 7.091 ns
33/100
   Vector  523.585 ns => 141.951 ns
BitVector  569.180 ns => 142.770 ns
50/100
   Vector  524.358 ns => 230.508 ns
BitVector  568.878 ns => 236.570 ns
66/100
   Vector  522.423 ns => 146.162 ns
BitVector  566.854 ns => 148.120 ns
99/100
   Vector  520.846 ns => 17.125 ns
BitVector  567.708 ns => 9.160 ns
100/100
   Vector  525.170 ns => 7.607 ns
BitVector  565.469 ns => 2.916 ns
0/300
   Vector  1.603 μs => 11.892 ns
BitVector  1.734 μs => 3.118 ns
1/300
   Vector  1.591 μs => 22.482 ns
BitVector  1.734 μs => 9.123 ns
100/300
   Vector  1.603 μs => 386.667 ns
BitVector  1.724 μs => 431.571 ns
150/300
   Vector  1.578 μs => 641.667 ns
BitVector  1.745 μs => 686.500 ns
200/300
   Vector  1.620 μs => 386.824 ns
BitVector  1.731 μs => 409.529 ns
299/300
   Vector  1.603 μs => 22.510 ns
BitVector  1.719 μs => 10.790 ns
300/300
   Vector  1.613 μs => 11.894 ns
BitVector  1.747 μs => 3.124 ns
0/1000
   Vector  5.050 μs => 26.393 ns
BitVector  5.425 μs => 4.092 ns
1/1000
   Vector  5.042 μs => 51.202 ns
BitVector  5.467 μs => 9.602 ns
333/1000
   Vector  5.050 μs => 1.282 μs
BitVector  5.483 μs => 1.349 μs
500/1000
   Vector  5.042 μs => 2.188 μs
BitVector  5.552 μs => 2.273 μs
666/1000
   Vector  5.067 μs => 1.284 μs
BitVector  5.467 μs => 1.353 μs
999/1000
   Vector  5.025 μs => 51.367 ns
BitVector  5.500 μs => 13.436 ns
1000/1000
   Vector  5.050 μs => 26.428 ns
BitVector  5.467 μs => 4.095 ns
0/3000
   Vector  16.167 μs => 73.308 ns
BitVector  17.458 μs => 9.461 ns
1/3000
   Vector  17.250 μs => 122.350 ns
BitVector  17.500 μs => 17.718 ns
1000/3000
   Vector  16.375 μs => 3.768 μs
BitVector  17.541 μs => 3.986 μs
1500/3000
   Vector  15.917 μs => 6.250 μs
BitVector  17.500 μs => 6.781 μs
2000/3000
   Vector  16.167 μs => 3.762 μs
BitVector  17.458 μs => 4.012 μs
2999/3000
   Vector  16.375 μs => 122.225 ns
BitVector  17.584 μs => 21.094 ns
3000/3000
   Vector  16.125 μs => 73.357 ns
BitVector  17.542 μs => 9.229 ns
0/10000
   Vector  55.541 μs => 264.134 ns
BitVector  59.750 μs => 25.198 ns
1/10000
   Vector  55.542 μs => 380.872 ns
BitVector  59.792 μs => 51.818 ns
3333/10000
   Vector  55.833 μs => 12.896 μs
BitVector  59.625 μs => 13.271 μs
5000/10000
   Vector  55.792 μs => 21.542 μs
BitVector  59.916 μs => 22.458 μs
6666/10000
   Vector  55.791 μs => 12.896 μs
BitVector  59.833 μs => 13.333 μs
9999/10000
   Vector  55.916 μs => 381.026 ns
BitVector  59.584 μs => 52.181 ns
10000/10000
   Vector  55.833 μs => 264.018 ns
BitVector  59.833 μs => 25.227 ns
0/100000
   Vector  568.625 μs => 2.667 μs
BitVector  614.542 μs => 312.946 ns
1/100000
   Vector  568.834 μs => 3.863 μs
BitVector  614.167 μs => 472.889 ns
33333/100000
   Vector  567.417 μs => 126.042 μs
BitVector  616.000 μs => 134.125 μs
50000/100000
   Vector  569.167 μs => 210.791 μs
BitVector  613.417 μs => 228.792 μs
66666/100000
   Vector  568.875 μs => 130.125 μs
BitVector  616.000 μs => 133.792 μs
99999/100000
   Vector  567.708 μs => 3.682 μs
BitVector  610.458 μs => 459.641 ns
100000/100000
   Vector  567.167 μs => 2.737 μs
BitVector  613.375 μs => 321.756 ns
0/10000000
   Vector  63.672 ms => 279.417 μs
BitVector  68.229 ms => 33.416 μs
1/10000000
   Vector  62.899 ms => 330.542 μs
BitVector  70.316 ms => 44.708 μs
3333333/10000000
   Vector  64.309 ms => 15.644 ms
BitVector  69.985 ms => 14.021 ms
5000000/10000000
   Vector  63.248 ms => 26.143 ms
BitVector  68.448 ms => 24.244 ms
6666666/10000000
   Vector  63.256 ms => 16.225 ms
BitVector  67.988 ms => 14.093 ms
9999999/10000000
   Vector  62.998 ms => 374.500 μs
BitVector  69.414 ms => 44.875 μs
10000000/10000000
   Vector  64.570 ms => 278.209 μs
BitVector  68.048 ms => 34.375 μs
0/100000000
   Vector  1.140 s (without a warmup) => 2.830 ms
BitVector  686.314 ms (without a warmup) => 353.542 μs
1/100000000
   Vector  1.122 s (without a warmup) => 3.850 ms
BitVector  680.261 ms (without a warmup) => 467.875 μs
33333333/100000000
   Vector  1.120 s (without a warmup) => 533.287 ms (without a warmup)
BitVector  665.777 ms (without a warmup) => 155.222 ms
50000000/100000000
   Vector  1.117 s (without a warmup) => 907.304 ms (without a warmup)
BitVector  693.189 ms (without a warmup) => 260.482 ms (without a warmup)
66666666/100000000
   Vector  1.114 s (without a warmup) => 533.780 ms (without a warmup)
BitVector  700.392 ms (without a warmup) => 154.736 ms
99999999/100000000
   Vector  1.114 s (without a warmup) => 4.231 ms
BitVector  677.860 ms (without a warmup) => 992.208 μs
100000000/100000000
   Vector  1.116 s (without a warmup) => 2.803 ms
BitVector  696.462 ms (without a warmup) => 347.667 μs
 49.686778 seconds (16.52 M allocations: 9.619 GiB, 1.54% gc time, 0.84% compilation time: 10% of which was recompilation)
```