@jakobnissen 2023-11-14
Maybe this is a dumb idea, but can't this be done in linear time? Suppose we have an array of length 10 with 4 trues and 6 falses.
Then, for the first element, set it true with probability 0.4. Then, update the probability for the second element based on whether the first was true or false: If true, it's now 3/9, else it's 4/9. I believe this is unbiased, but any implementation would need to be careful about float rounding issues.