@barucden 2023-11-24
The method is defined for `AbstractArray`s, so I am wondering is `length(a)` always cheap? If not, maybe it is worth an extra line to avoid two `length` calls?